+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Bugfixes
 
 - [#5924](https://github.com/influxdata/influxdb/issues/5924): Missing data after using influx\_tsm
+- [#5594](https://github.com/influxdata/influxdb/pull/5594): Fix missing url params on lease redirect - @oldmantaiter
 
 ## v0.10.2 [2016-03-03]
 ### Bugfixes