+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.12.0 [unreleased]
+
+### Features
+
+- [#5960](https://github.com/influxdata/influxdb/pull/5960): Provide ability to delete a shard via the Meta Client.
+
 ## v0.11.0 [unreleased]
 
 ### Features